SQLiteストアを機能別モジュールへ分割 / Split SQLite store into feature modules
